--- CPP/Neural_Simulator/Neural_Simulator.cpp
@@ -1,10 +1,0 @@
-#include <iostream>
-#include "Models/Log.h"
-
-int main()
-{
-    Log::SetLevel(LogInfo);
-    Log::DisplayInfo("Hello world!");
-    std::cin.get();
-}
-
