--- CPP/Neural_Simulator/Neural_Simulator.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.1.32414.318
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Neural_Simulator", "Neural_Simulator.vcxproj", "{034A6A2B-FBFC-415E-9342-3642EF2E3A80}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Debug|x64.ActiveCfg = Debug|x64
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Debug|x64.Build.0 = Debug|x64
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Debug|x86.ActiveCfg = Debug|Win32
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Debug|x86.Build.0 = Debug|Win32
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Release|x64.ActiveCfg = Release|x64
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Release|x64.Build.0 = Release|x64
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Release|x86.ActiveCfg = Release|Win32
-		{034A6A2B-FBFC-415E-9342-3642EF2E3A80}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5D07C725-7165-4EF6-9FF1-8C4DC7583937}
-	EndGlobalSection
-EndGlobal
