--- CPP/Neural_Simulator/x64/Debug/Neural_Simulator.exe.recipe
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>
-    <ProjectOutput>
-      <FullPath>D:\Repos\Personal repos\Izhikevich_neural_simulator\CPP\Neural_Simulator\x64\Debug\Neural_Simulator.exe</FullPath>
-    </ProjectOutput>
-  </ProjectOutputs>
-  <ContentFiles />
-  <SatelliteDlls />
-  <NonRecipeFileRefs />
-</Project>
