--- CPP/Neural_Simulator/x64/Debug/Neural_Simulator.log
@@ -1,5 +1,0 @@
-﻿  Log.cpp
-  Neuron.cpp
-  Neural_Simulator.cpp
-  Generando código...
-  Neural_Simulator.vcxproj -> D:\Repos\Personal repos\Izhikevich_neural_simulator\CPP\Neural_Simulator\x64\Debug\Neural_Simulator.exe
