--- CPP/Neural_Simulator/x64/Debug/Neural_Simulator.tlog/Neural_Simulator.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.31.31103:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|D:\Repos\Personal repos\Izhikevich_neural_simulator\CPP\Neural_Simulator\|
