--- CPP/Neural_Simulator/x64/Debug/Neural_Simulator.vcxproj.FileListAbsolute.txt
@@ -1,1 +1,0 @@
-D:\Repos\Personal repos\Izhikevich_neural_simulator\CPP\Neural_Simulator\x64\Debug\Neural_Simulator.exe
