+++ setup.cfg
@@ -1,0 +1,23 @@
+[metadata]
+name = brainx
+description = Neural simulation module
+author = Mario González
+license = MIT
+license_file = LICENSE
+platforms = unix, linux, osx, cygwin, win32
+classifiers =
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+
+[options]
+packages = 
+    brainx
+install_requires = 
+    requests>=2
+python_requires = >= 3.8
+package_dir =
+    =src
+zip_safe = no
