+++ setup.py
@@ -1,0 +1,4 @@
+from setuptools import setup
+
+if __name__ == "__main__" :
+    setup()
